requirements.txt
# Django
asgiref==3.2.3
Django==3.0.2
pytz==2019.3
sqlparse==0.3.0

# Python Decouple
python-decouple==3.3

# dj-database-url
dj-database-url==0.5.0

# Pillow
Pillow==7.0.0
